+++ build.js
@@ -2,9 +2,27 @@ import { build } from 'esbuild';
 import { lessLoader } from "esbuild-plugin-less";
 import progress from "@olton/esbuild-plugin-progress";
 import autoprefixer from "@olton/esbuild-plugin-autoprefixer";
+import unlink from "@olton/esbuild-plugin-unlink";
 import { replace } from "esbuild-plugin-replace";
+import pkg from "./package.json" assert {type: "json"};
 
 const production = process.env.MODE === "production"
+const version = pkg.version
+
+const banner = `
+/*!
+ ███╗   ███╗███████╗████████╗██████╗  ██████╗     ██╗   ██╗██╗
+ ████╗ ████║██╔════╝╚══██╔══╝██╔══██╗██╔═══██╗    ██║   ██║██║
+ ██╔████╔██║█████╗     ██║   ██████╔╝██║   ██║    ██║   ██║██║
+ ██║╚██╔╝██║██╔══╝     ██║   ██╔══██╗██║   ██║    ██║   ██║██║
+ ██║ ╚═╝ ██║███████╗   ██║   ██║  ██║╚██████╔╝    ╚██████╔╝██║
+ ╚═╝     ╚═╝╚══════╝   ╚═╝   ╚═╝  ╚═╝ ╚═════╝      ╚═════╝ ╚═╝                                                             
+
+ * Metro UI v${version} Components Library  (https://metroui.org.ua)
+ * Copyright 2012-${new Date().getFullYear()} by Serhii Pimenov
+ * Licensed under MIT
+ !*/
+`
 
 await build({
     entryPoints: ['./source/default.js'],
@@ -12,6 +30,9 @@ await build({
     bundle: true,
     minify: production,
     sourcemap: false,
+    banner: {
+        js: banner
+    },
     plugins: [
         progress({
             text: 'Building Metro UI...',
@@ -20,7 +41,8 @@ await build({
         lessLoader(),
         autoprefixer(),
         replace({
-            '__BUILD_TIME__': new Date().toLocaleString()
+            '__BUILD_TIME__': new Date().toLocaleString(),
+            '__VERSION__': version,
         })
     ],
 })
@@ -31,6 +53,9 @@ await build({
     bundle: true,
     minify: production,
     sourcemap: false,
+    banner: {
+        js: banner
+    },
     plugins: [
         progress({
             text: 'Building Metro UI with icons...',
@@ -39,7 +64,8 @@ await build({
         lessLoader(),
         autoprefixer(),
         replace({
-            '__BUILD_TIME__': new Date().toLocaleString()
+            '__BUILD_TIME__': new Date().toLocaleString(),
+            '__VERSION__': version,
         })
     ],
 })
@@ -56,8 +82,8 @@ await build({
             succeedText: 'Metro UI icons built successfully in %s ms!'
         }),
         lessLoader(),
-        replace({
-            '__BUILD_TIME__': new Date().toLocaleString()
+        unlink({
+            files: ['./lib/icons.js']
         })
     ],
 })