+++ source/components/navview/navview.js
@@ -97,6 +97,12 @@
                 element.addClass("compacted handmade");
             }
 
+            const state = Metro.storage.getItem("navview-compacted");
+            if (state === true) {
+                element.removeClass("expanded");
+                element.addClass("compacted handmade");
+            }
+
             pane = element.children(".navview-pane");
             content = element.children(".navview-content");
             toggle = $(o.toggle);
@@ -168,12 +174,16 @@
 
             if (this.paneToggle !== null) {
                 this.paneToggle.on(Metro.events.click, function(){
-                    that.pane.toggleClass("open");
+                    //that.pane.toggleClass("open");
                 })
             }
 
             $(globalThis).on(Metro.events.resize, () => {
                 this._recalc();
+
+                if (o.saveState === true && element.hasClass("compacted")) {
+                    return false;
+                }
 
                 if (!element.hasClass("handmade")) {
                     if (Metro.utils.isValue(o.expandPoint) && Metro.utils.mediaExist(o.expandPoint)) {
@@ -196,8 +206,10 @@
             element.toggleClass("handmade");
 
             if (element.hasClass("compacted")) {
+                Metro.storage.setItem("navview-compacted", true);
                 Metro.utils.exec(o.onPaneClose, null, this)
             } else {
+                Metro.storage.setItem("navview-compacted", false);
                 Metro.utils.exec(o.onPaneOpen, null, this)
             }
         },
@@ -227,6 +239,13 @@
             const element = this.element, o = this.options;
             element.addClass("compacted handmade")
             element.removeClass("expanded")
+            this._recalc()
+        },
+
+        expand: function(){
+            const element = this.element, o = this.options;
+            element.addClass("expanded")
+            element.removeClass("compacted handmade")
             this._recalc()
         },