+++ source/core/index.js
@@ -1,3 +1,3 @@
-import "../string"
+import "../string/index.js"
 import "./global.js"
 import "./metro.js"